permission handling

+++ app/src/main/java/com/rachitbhutani/allesample/share/ShareFragment.kt
@@ -1,6 +1,13 @@
 package com.rachitbhutani.allesample.share
 
+import android.Manifest.permission.READ_EXTERNAL_STORAGE
+import android.Manifest.permission.READ_MEDIA_IMAGES
+import android.Manifest.permission.WRITE_EXTERNAL_STORAGE
+import android.app.RecoverableSecurityException
 import android.content.Intent
+import android.content.IntentSender
+import android.content.pm.PackageManager
+import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 import android.provider.MediaStore
@@ -8,6 +15,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
+import android.widget.Toast
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.annotation.RequiresApi
 import androidx.core.content.ContextCompat
 import androidx.core.view.children
 import androidx.core.view.isVisible
@@ -18,6 +28,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.rachitbhutani.allesample.MainActivity
 import com.rachitbhutani.allesample.MainActivity.Companion.DELETE_ITEM_REQUEST_CODE
 import com.rachitbhutani.allesample.MainViewModel
 import com.rachitbhutani.allesample.R
@@ -26,7 +37,8 @@ import com.rachitbhutani.allesample.databinding.LabelItemBinding
 import com.rachitbhutani.allesample.share.model.ScreenshotItem
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
-import kotlinx.coroutines.launch
+import java.io.File
+
 
 @AndroidEntryPoint
 class ShareFragment : Fragment(), ScreenshotListCallback {
@@ -34,9 +46,16 @@ class ShareFragment : Fragment(), ScreenshotListCallback {
     private lateinit var binding: FragmentShareBinding
     private lateinit var adapter: ScreenshotListAdapter
 
-    var deletingAtPos: Int? = null
+    private var deletingAtPos: Int? = null
 
     private val viewModel: MainViewModel by activityViewModels()
+
+    private val requestPermissions =
+        registerForActivityResult(ActivityResultContracts.RequestMultiplePermissions()) { isGranted ->
+            if (isGranted[READ_MEDIA_IMAGES] == true || isGranted[READ_EXTERNAL_STORAGE] == true) loadImages()
+            else if (isGranted[WRITE_EXTERNAL_STORAGE] == true) deleteActiveScreenshot()
+            else Toast.makeText(requireContext(), "Permission denied", Toast.LENGTH_SHORT).show()
+        }
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -52,12 +71,39 @@ class ShareFragment : Fragment(), ScreenshotListCallback {
         setupUi()
         setupObservers()
         setupListeners()
+        if (checkAndRequestReadPermissions()) loadImages()
+    }
+
+    private fun checkAndRequestReadPermissions(): Boolean {
+
+        val requiredPermission =
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
+                READ_MEDIA_IMAGES
+            } else {
+                READ_EXTERNAL_STORAGE
+            }
+        return if (ContextCompat.checkSelfPermission(
+                requireContext(),
+                requiredPermission
+            ) == PackageManager.PERMISSION_GRANTED
+        ) {
+            true
+        } else {
+            requestPermissions.launch(arrayOf(requiredPermission))
+            false
+        }
     }
 
     private fun setupListeners() {
         binding.run {
             tvDelete.setOnClickListener {
-                deleteActiveScreenshot()
+                if (ContextCompat.checkSelfPermission(
+                        requireContext(),
+                        WRITE_EXTERNAL_STORAGE
+                    ) == PackageManager.PERMISSION_GRANTED
+                )
+                    deleteActiveScreenshot()
+                else requestPermissions.launch(arrayOf(WRITE_EXTERNAL_STORAGE))
             }
             tvInfo.setOnClickListener {
                 if (llCollections.isVisible) {
@@ -83,7 +129,7 @@ class ShareFragment : Fragment(), ScreenshotListCallback {
             requireActivity().startActivity(
                 Intent.createChooser(
                     sharingIntent,
-                    "Share the screenshot"
+                    getString(R.string.share_the_screenshot)
                 )
             )
         }
@@ -93,38 +139,61 @@ class ShareFragment : Fragment(), ScreenshotListCallback {
         viewModel.activeScreenshotLiveData.value?.let {
             deletingAtPos = adapter.snapshot().indexOf(it)
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-                val intent = MediaStore.createTrashRequest(
-                    requireActivity().contentResolver,
-                    listOf(it.uri),
-                    true
-                )
-                requireActivity().startIntentSenderForResult(
-                    intent.intentSender,
-                    DELETE_ITEM_REQUEST_CODE,
-                    null,
-                    0,
-                    0,
-                    0
-                )
-            } else {
-                it.uri?.let { uri ->
-                    requireActivity().contentResolver.delete(
-                        uri,
-                        null,
-                        null
-                    )
+                trashScreenshot(it.uri)
+            } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+                it.path?.let { path ->
+                    val file = File(path)
+                    if (file.exists()) {
+                        try {
+                            // Perform the deletion operation
+                            requireActivity().contentResolver.delete(it.uri!!, null, null)
+                            deletingAtPos?.let { pos -> adapter.notifyItemRemoved(pos) }
+                            viewModel.refreshData()
+                        } catch (e: RecoverableSecurityException) {
+                            val intentSender: IntentSender = e.userAction.actionIntent.intentSender
+                            startIntentSenderForResult(
+                                intentSender,
+                                MainActivity.DELETE_EXCEPTION_ALLOWED,
+                                null,
+                                0,
+                                0,
+                                0,
+                                null
+                            )
+                        }
+                    }
                 }
             }
         }
 
     }
 
-    private fun setupObservers() {
-        lifecycleScope.launch {
+    @RequiresApi(Build.VERSION_CODES.R)
+    private fun trashScreenshot(uri: Uri?) {
+        val intent = MediaStore.createTrashRequest(
+            requireActivity().contentResolver,
+            listOf(uri),
+            true
+        )
+        requireActivity().startIntentSenderForResult(
+            intent.intentSender,
+            DELETE_ITEM_REQUEST_CODE,
+            null,
+            0,
+            0,
+            0
+        )
+    }
+
+    private fun loadImages() {
+        viewLifecycleOwner.lifecycleScope.launchWhenCreated {
             viewModel.galleryData.collectLatest {
                 adapter.submitData(it)
             }
         }
+    }
+
+    private fun setupObservers() {
         viewModel.activeScreenshotLiveData.observe(viewLifecycleOwner) {
             deletingAtPos?.let { pos ->
                 adapter.notifyItemRemoved(pos)
